+++ CHANGELOG.md
@@ -16,6 +16,74 @@ individual release will not have many changes within it. Below is a list
 of the releases we've made so far, along with what was changed within
 each release.
 
+## 0.13.0 (2026-05-04)
+
+## 0.12.0 (2026-04-30)
+
+### Feat
+
+- ✨ beautify CLI output in docs (#31)
+
+## 0.11.0 (2026-04-24)
+
+### Feat
+
+- ✨ add possibility to return example path (#29)
+
+## 0.10.0 (2026-04-24)
+
+### Feat
+
+- ✨ improve example I/O syntax (#26)
+
+## 0.9.0 (2026-04-23)
+
+### Feat
+
+- ✨ add woolly example package (#28)
+
+## 0.8.0 (2026-04-23)
+
+### Feat
+
+- ✨ add `write_properties()` (#27)
+
+## 0.7.1 (2026-04-21)
+
+### Fix
+
+- 🐛 export example name enum (#25)
+
+## 0.7.0 (2026-04-20)
+
+### Feat
+
+- ✨ allow config file to be optional (#23)
+
+## 0.6.0 (2026-04-17)
+
+### Feat
+
+- ✨ `read_example_datapackage()` and (re)organize examples (#15)
+
+## 0.5.1 (2026-04-17)
+
+### Refactor
+
+- ♻️ use descriptive return type (#21)
+
+## 0.5.0 (2026-04-15)
+
+### Feat
+
+- ✨ add example datapackage jsons (#10)
+
+## 0.4.0 (2026-04-14)
+
+### Feat
+
+- :sparkles: add functional helpers (#9)
+
 ## 0.3.0 (2026-04-13)
 
 ### Feat